add Movie type to types.ts
